requirements.txt
openai-whisper 
PyAudio 
python-multipart 
flask
flask_cors
dotenv
google-genai
gtts